Code Lint, Notifications when signed in, Notifications fix
BACKGROUND.JS:
- Fixed the "wrongToken" when closing tabs. It's set to "null" in order
to be recognized as invalid token. This way, notifications only come up
when user is signed IN.

- Listener for accessToken is commented out. Since it comes through
LocalStorage, there's no need for message. Testing for stability.

OPTIONS.JS
- Removed a lot of legacy code on notifications.
- Removed passAccessToken function, as it already goes through
LocalStorage
- Onload, OAuth and "if is valid" gotten a call to update badge. This
way, Sign In now updates the badge for colored and counting tabs when
clicked.
- chrome.storage.local.set removed.

HELPERS.JS
- UpdateBadge now checks localStorage directly after accessToken,
instead of Logged. It also evaluates inside the function.
- Tabs instead of "  "
- Commented out the identity flow for sign out.