Injected menus and pomodoro corrections
- Lots of properties are defined in the CSS to minimize collision. Some
properties are defined in DOM level and it's simply not possible to make
it look the same across every site.

BACKGROUND.js
- Revised structure for check blacklist, including a check for
pomoFocus.active as a parent for locked and daily.specialList
- Trying to delete dailyPomo and lockZap in every finish.countdown
event. But it just doesn't work.
- calibratePomoFocus function checks if it should be active according to
end time. If not, it resets audio, daily, specialList, instaZap, lockMe
and dailyPomo. This fixes the fact the variables weren't being reset at
the pomo ends
- Last fix also fixes special list and lock me getting stuck after
pomodoros.
- Fixed typo for tabNumbersActive check

OPTIONS.HTML
- Daily pomos now have nearly the same duration possiblities as regular
tasks