Creation and relocation of todoist Properties and methods
- originalTodo property gets the regular testTodo methods create, update
and delete. They can be changed for a method that will trigger todoist
functions when called, in order to integrate todoist and Pavlok.

- timeouts property set to control when the background will try to get
tasks from Todoist.
- intervalChecks property set how many seconds between automated checks.
Default is 30

HELPERS
- seq_no gets the latest seq_no from the latest todoist call
- readImportedTasks returns an array of tasks obtained from the latest
todoist call
- taskExternalId returns either the task with the given external ID or
null. Used to tell which task should get updated on Pavlok side when
updates come via getTasks
- addToDoListeners trades the original testTodo backend functions for
methods which have a call for the todoist functions.
- random function corrected to give the proper length of string
(requested in the function call) instead of a fixed 5

BACKEND
- getToken function activates addToDoListeners if successfull
- sendTask updates the sent tasks on Pavlok's side if successful
- complete, uncomplete and delete use the proper externalId, instead of
Pavlok's local id
- import2 became import
- import2 checks if tasks exists. If so, it updates existing task. If
not, it creates a new one.

- Conditional addToDoListeners on background start if there's an access
token for todoist in the local storage

CLEANING
- import method deleted
- export method deleted
- onTheList empty function deleted

BACKGROUND
- Listeners added for changes in tasks, calling the new methods of
todoist
- Naming adjusted for the new namespaces (backend, frontend, helpers)
- Reacting to ToDo calls

TODO
- new task creation (regular and daily), delete,  done checkbox clicks
now messages background to make logic happen there and interface on
popup