Popup blocker
If blacklist zaps go over 3 per minute, you get a popup asking what you
are doing. That popup will prevent stimuli from being sent when the user
is not using the pavlok (chances are, most people receiving that many
stimuli in a minute are either in a training session, so they comply or
not using pavlok, so they shouldn't be getting stimuli).