Blacklist interrupting popups
If a user gets more than n (3) zaps in a minute from the blacklist, a
popup appears asking what the user is doing.

This popup prevent more stimuli from being sent and tabs change. This is
our best effort to avoid people not wearing Pavlok, but with an active
black list, from sending thousands of stimuli per hour to our databases.